refactor: extract TypeSummaryCards component
